rename AudioFile/{BaseAudioFile => base}.hpp
