+++ include/AudioFile/base.hpp
